Featureコンポーネントのstoreの取得方法変更とcareerコンポーネントをFeatureコンポーネントに統合
