--- frontend/src/components/Career.vue
@@ -1,31 +1,0 @@
-<script setup>
-import { onMounted } from 'vue';
-import { useCareerStore } from '@/stores/careerStore';
-
-const careerStore = useCareerStore();
-
-onMounted(async () => {
-  await careerStore.fetchCareers();
-});
-</script>
-
-<template>
-  <div id="career" class="m-1">
-    <h3 class="text-2xl">Career</h3>
-    <div v-if="careerStore.error" class="m-1">
-      <p>Failed to fetch careers.</p>
-    </div>
-    <div v-else-if="careerStore.loading" class="m-1"><p>Loading...</p></div>
-    <ul v-else class="list-none m-1">
-      <li class="flex" v-for="career in careerStore.careers" :key="career.id">
-        <span class="w-40"
-          >{{ career.period_from }}~{{ career.period_to ?? 'current' }}</span
-        >
-        <span class="w-5">:</span>
-        <span class="w-35">{{ career.position }}</span>
-      </li>
-    </ul>
-  </div>
-</template>
-
-<style></style>
