+++ frontend/src/components/Feature.vue
@@ -1,21 +1,39 @@
 <script setup>
 import { onMounted } from 'vue';
+import { useCareerStore } from '@/stores/careerStore';
 import { useSkillStore } from '@/stores/skillStore';
 import { useHobbyStore } from '@/stores/hobbyStore';
 import { useCertificationStore } from '@/stores/certificationStore';
 
+const careerStore = useCareerStore();
 const skillStore = useSkillStore();
 const hobbyStore = useHobbyStore();
 const certificationStore = useCertificationStore();
 
-onMounted(async () => {
-  await skillStore.fetchSkills();
-  await hobbyStore.fetchHobbies();
-  await certificationStore.fetchCertifications();
-});
+careerStore.fetchCareers();
+skillStore.fetchSkills();
+hobbyStore.fetchHobbies();
+certificationStore.fetchCertifications();
 </script>
 
 <template>
+  <div id="career" class="m-1">
+    <h3 class="text-2xl">Career</h3>
+    <div v-if="careerStore.error" class="m-1">
+      <p>Failed to fetch careers.</p>
+    </div>
+    <div v-else-if="careerStore.loading" class="m-1"><p>Loading...</p></div>
+    <ul v-else class="list-none m-1">
+      <li class="flex" v-for="career in careerStore.careers" :key="career.id">
+        <span class="w-40"
+          >{{ career.period_from }}~{{ career.period_to ?? 'current' }}</span
+        >
+        <span class="w-5">:</span>
+        <span class="w-35">{{ career.position }}</span>
+      </li>
+    </ul>
+  </div>
+
   <div class="m-1">
     <h3 class="text-2xl">Skills</h3>
     <div v-if="skillStore.error" class="m-1">
